requirements.txt
django==3.2.12
django-filer==2.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
whitenoise==6.0.0
sentry-sdk==1.5.8
beautifulsoup4==4.10.0
requests==2.27.1
pytube==12.0.0